requirements.txt
requests
beautifulsoup4
twilio
google-generativeai
python-dotenv
